Move scanner common code to common folder
