+++ common/reconciliation/invariants/historyExists.go
@@ -27,7 +27,7 @@ import (
 
 	c "github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/persistence"
-	"github.com/uber/cadence/service/worker/scanner/executions/common"
+	"github.com/uber/cadence/common/reconciliation/common"
 )
 
 const (
