+++ common/reconciliation/invariants/util.go
@@ -23,7 +23,7 @@
 package invariants
 
 import (
-	"github.com/uber/cadence/service/worker/scanner/executions/common"
+	"github.com/uber/cadence/common/reconciliation/common"
 )
 
 func checkBeforeFix(
