+++ service/worker/scanner/executions/activities.go
@@ -33,7 +33,7 @@ import (
 
 	c "github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/metrics"
-	"github.com/uber/cadence/service/worker/scanner/executions/common"
+	"github.com/uber/cadence/common/reconciliation/common"
 	"github.com/uber/cadence/service/worker/scanner/executions/shard"
 )
