+++ service/worker/scanner/executions/aggregators.go
@@ -25,7 +25,7 @@ package executions
 import (
 	"fmt"
 
-	"github.com/uber/cadence/service/worker/scanner/executions/common"
+	"github.com/uber/cadence/common/reconciliation/common"
 )
 
 type (
