+++ service/worker/scanner/executions/workflows.go
@@ -28,8 +28,8 @@ import (
 	"go.uber.org/cadence/workflow"
 
 	"github.com/uber/cadence/common/metrics"
+	"github.com/uber/cadence/common/reconciliation/common"
 	"github.com/uber/cadence/common/resource"
-	"github.com/uber/cadence/service/worker/scanner/executions/common"
 )
 
 const (
